+++ packages/strapi-admin/admin/src/components/LeftMenuFooter/messages.json
@@ -1,12 +1,4 @@
 {
-  "documentation": {
-    "id": "app.components.LeftMenuFooter.documentation",
-    "defaultMessage": "Documentation"
-  },
-  "help": {
-    "id": "app.components.LeftMenuFooter.help",
-    "defaultMessage": "Help"
-  },
   "poweredBy": {
     "id": "app.components.LeftMenuFooter.poweredBy",
     "defaultMessage": "Proudly powered by "
