+++ packages/strapi-admin/admin/src/components/StaticLinks/StyledLink.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import { Link } from 'react-router-dom';
+
+const StyledLink = styled(Link)``;
+
+export default StyledLink;
